# src/swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Swagger Todo App - OpenAPI 3.0",
      "description": "This is a API docs of todo application \\\nAuthor: Hai (Mink)",
      "contact": {
        "email": "nguyen.van.hai02@azoom.jp"
      },
      "version": "1.0.0"
    },
    "servers": [
      {
        "url": "http://localhost:3000"
      }
    ],
...
    "paths": {
...
      "/todos/{todoId}": {
        "get": {
          "tags": [
            "Todo"
          ],
          "summary": "Find todo by ID",
          "description": "Returns a single todo",
          "operationId": "getTodoById",
          "parameters": [
            {
              "name": "todoId",
              "in": "path",
              "description": "ID of todo to return",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/Todo"
                  }
                },
                "application/xml": {
                  "schema": {
                    "$ref": "#/components/schemas/Todo"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid ID supplied"
            },
            "404": {
              "description": "Todo not found"
            }
          }
        },
        "put": {
          "tags": [
            "Todo"
          ],
          "summary": "Updates a todo in the store with form data",
          "description": "",
          "operationId": "updateTodoWithForm",
          "parameters": [
            {
              "name": "todoId",
              "in": "path",
              "description": "ID of todo that needs to be updated",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "requestBody": {
            "description": "Create a new todo in the store",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              },
              "application/x-www-form-urlencoded": {
                "schema": {
                  "$ref": "#/components/schemas/Todo"
                }
              }
            },
            "required": true
          },
          "responses": {
            "200":{
                "description": "Created successfully",
                "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ExistingTodo"
                  }
                },
                "application/xml": {
                  "schema": {
                    "$ref": "#/components/schemas/ExistingTodo"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input"
            }
          }
        },
        "delete": {
          "tags": [
            "Todo"
          ],
          "summary": "Deletes a todo",
          "description": "delete a todo",
          "operationId": "deleteTodo",
          "parameters": [
            {
              "name": "todoId",
              "in": "path",
              "description": "Todo id to delete",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int64"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Delete sucessfully"
            },
            "400": {
              "description": "Invalid todo value"
            }
          }
        }
      }
    },
    "components": {
      "schemas": {
        "Todo": {
          "required": [
            "description",
            "status"
          ],
          "type": "object",
          "properties": {
            "description": {
              "type": "string",
              "example": "Todo do something"
            },
            "status": {
              "type": "integer",
              "description": "todo status are done, todo (not done), in [0, 1]",
              "enum": [
                0,
                1
              ]
            }
          },
          "xml": {
            "name": "todo"
          }
        },
        "ExistingTodo": {
            
          "required": [
            "description",
            "status"
          ],
          "type": "object",
          "properties": {
            "id":{
                "type": "integer",
                "format": "int64",
                "example": 10
            },
            "description": {
              "type": "string",
              "example": "Todo do something"
            },
            "status": {
              "type": "integer",
              "description": "todo status are done, todo (not done), in [0, 1]",
              "enum": [
                0,
                1
              ]
            }
          },
          "xml": {
            "name": "todo"
          }
        }
      },
...
    }
  }